@@ -1,7 +1,8 @@
 {
     "name": "@mkody/twitch-emoticons",
-    "version": "2.9.5",
+    "version": "3.0.0-beta.1",
     "description": "Gets Twitch, BTTV, FFZ and 7TV emotes as well as parsing text to emotes!",
+    "type": "module",
     "main": "src/index.js",
     "typings": "typings/index.d.ts",
     "keywords": [
@@ -26,14 +27,15 @@
     ],
     "license": "MIT",
     "dependencies": {
-        "@twurple/api": "^7.4.0",
-        "@twurple/auth": "^7.4.0",
+        "@twurple/api": "^8.0.3",
+        "@twurple/auth": "^8.0.3",
         "axios": "^1.13.2"
     },
     "devDependencies": {
         "@eslint/js": "^9.39.2",
         "@jest/globals": "^30.2.0",
         "@microsoft/eslint-formatter-sarif": "^3.1.0",
+        "cross-env": "^10.1.0",
         "docdash": "^2.0.2",
         "esbuild": "^0.27.2",
         "eslint": "^9.39.2",
@@ -47,7 +49,7 @@
         "build": "esbuild ./src/index.js --bundle --minify --keep-names --sourcemap --format=esm --outfile=./dist/TwitchEmoticons.min.js",
         "docs": "jsdoc --configure .jsdoc.json",
         "lint": "eslint ./src ./test",
-        "test": "jest"
+        "test": "cross-env NODE_OPTIONS=\"--experimental-vm-modules\" jest"
     },
     "repository": {
         "type": "git",
@@ -62,8 +64,8 @@
         "test": "test"
     },
     "engines": {
-        "node": ">=18.0.0",
-        "pnpm": ">=9"
+        "node": ">=20.19.0",
+        "pnpm": ">=10"
     },
     "packageManager": "pnpm@10.28.1",
     "pnpm": {
