@@ -84,4 +84,4 @@ class Emote {
     }
 }
 
-module.exports = Emote;
+export default Emote;
