@@ -1,5 +1,5 @@
-const Emote = require('./Emote');
-const Constants = require('../util/Constants');
+import Emote from './Emote.js';
+import Constants from '../util/Constants.js';
 
 /** @augments Emote */
 class FFZEmote extends Emote {
@@ -110,4 +110,4 @@ class FFZEmote extends Emote {
     }
 }
 
-module.exports = FFZEmote;
+export default FFZEmote;
