@@ -1,5 +1,5 @@
-const Emote = require('./Emote');
-const Constants = require('../util/Constants');
+import Emote from './Emote.js';
+import Constants from '../util/Constants.js';
 
 /** @augments Emote */
 class TwitchEmote extends Emote {
@@ -84,4 +84,4 @@ class TwitchEmote extends Emote {
     }
 }
 
-module.exports = TwitchEmote;
+export default TwitchEmote;
