@@ -76,4 +76,4 @@ class Collection extends Map {
     }
 }
 
-module.exports = Collection;
+export default Collection;
