@@ -3,8 +3,7 @@
  * https://jestjs.io/docs/configuration
  */
 
-/** @type {import('jest').Config} */
-const config = {
+export default {
     // All imported modules in your tests should be mocked automatically
     // automock: false,
 
@@ -194,5 +193,3 @@ const config = {
     // Whether to use watchman for file crawling
     // watchman: true
 };
-
-module.exports = config;
